feat(ts) make tsc happy
