+++ globals.native.d.ts
@@ -12,8 +12,21 @@ interface IWindow {
     JITSI_MEET_LITE_SDK: boolean;
     JitsiMeetJS: any;
     config: IConfig;
+    document: any;
+    innerHeight: number;
+    innerWidth: number;
     interfaceConfig: any;
     location: ILocation;
+    self: any;
+    top: any;
+
+    onerror: (event: string, source: any, lineno: any, colno: any, e: Error) => void;
+    onunhandledrejection: (event: any) => void;
+
+    setTimeout: typeof setTimeout;
+    clearTimeout: typeof clearTimeout;
+    setImmediate: typeof setImmediate;
+    clearImmediate: typeof clearImmediate;
 }
 
 interface INavigator {
@@ -22,6 +35,7 @@ interface INavigator {
 
 declare global {
     const APP: any;
+    const document: any;
     const interfaceConfig: any;
     const navigator: INavigator;
     const window: IWindow;
