+++ package.json
@@ -199,7 +199,7 @@
     "tsc:web": "tsc --noEmit --project tsconfig.web.json",
     "tsc:native": "tsc --noEmit --project tsconfig.native.json",
     "tsc:ci": "npm run tsc:web && npm run tsc:native",
-    "lint:ci": "eslint --ext .js,.ts,.tsx --max-warnings 0 . && npm run tsc:web",
+    "lint:ci": "eslint --ext .js,.ts,.tsx --max-warnings 0 . && npm run tsc:ci",
     "lang-sort": "./resources/lang-sort.sh",
     "lint-fix": "eslint --ext .js,.ts,.tsx --max-warnings 0 --fix .",
     "postinstall": "patch-package --error-on-fail && jetify",
