+++ react/features/analytics/handlers/AmplitudeHandler.ts
@@ -1,3 +1,5 @@
+/* eslint-disable lines-around-comment */
+
 import logger from '../logger';
 
 import AbstractHandler, { IEvent } from './AbstractHandler';
@@ -63,7 +65,7 @@ export default class AmplitudeHandler extends AbstractHandler {
      * @param {Object} userProps - The user portperties.
      * @returns {void}
      */
-    setUserProperties(userProps: Object) {
+    setUserProperties(userProps: any) {
         if (this._enabled) {
             amplitude.getInstance().setUserProperties(userProps);
         }
@@ -82,6 +84,7 @@ export default class AmplitudeHandler extends AbstractHandler {
             return;
         }
 
+        // @ts-ignore
         amplitude.getInstance().logEvent(this._extractName(event) ?? '', event);
     }
 
@@ -100,7 +103,9 @@ export default class AmplitudeHandler extends AbstractHandler {
 
         return {
             sessionId: amplitude.getInstance().getSessionId(),
+            // @ts-ignore
             deviceId: amplitude.getInstance().options.deviceId,
+            // @ts-ignore
             userId: amplitude.getInstance().options.userId
         };
     }
