+++ react/features/base/devices/functions.web.ts
@@ -5,10 +5,9 @@ import { ISettingsState } from '../settings/reducer';
 import { parseURLParams } from '../util/parseURLParams';
 
 import logger from './logger';
-import { IDevicesState } from './reducer';
+import { IDevicesState } from './types';
 
-
-declare const APP: any;
+export * from './functions.any';
 
 const webrtcKindToJitsiKindTranslator = {
     audioinput: 'audioInput',
@@ -238,24 +237,6 @@ export function getAudioOutputDeviceData(state: IReduxState) {
  */
 export function getVideoDeviceIds(state: IReduxState) {
     return state['features/base/devices'].availableDevices.videoInput?.map(({ deviceId }) => deviceId);
-}
-
-/**
- * Returns true if there are devices of a specific type or on native platform.
- *
- * @param {Object} state - The state of the application.
- * @param {string} type - The type of device: VideoOutput | audioOutput | audioInput.
- *
- * @returns {boolean}
- */
-export function hasAvailableDevices(state: IReduxState, type: string) {
-    if (state['features/base/devices'] === undefined) {
-        return true;
-    }
-
-    const availableDevices = state['features/base/devices'].availableDevices;
-
-    return Number(availableDevices[type as keyof typeof availableDevices]?.length) > 0;
 }
 
 /**
