--- react/features/base/devices/index.js
@@ -1,3 +1,0 @@
-export * from './actions';
-export * from './actionTypes';
-export * from './functions';
