+++ react/features/base/devices/types.ts
@@ -1,0 +1,17 @@
+/* eslint-disable lines-around-comment */
+
+export interface IDevicesState {
+    availableDevices: {
+        // @ts-ignore
+        audioInput?: MediaDeviceInfo[];
+        // @ts-ignore
+        audioOutput?: MediaDeviceInfo[];
+        // @ts-ignore
+        videoInput?: MediaDeviceInfo[];
+    };
+    pendingRequests: any[];
+    permissions: {
+        audio: boolean;
+        video: boolean;
+    };
+}
