+++ react/features/base/environment/checkChromeExtensionsInstalled.native.ts
@@ -1,0 +1,10 @@
+/**
+ * Checks whether the chrome extensions defined in the config file are installed or not.
+ *
+ * @param {Object} _config - Objects containing info about the configured extensions.
+ *
+ * @returns {Promise[]}
+ */
+export default function checkChromeExtensionsInstalled(_config: any = {}) {
+    return Promise.resolve([]);
+}
