--- react/features/base/ui/components/_.native.ts
@@ -1,1 +1,0 @@
-export * from './native';
