--- react/features/base/ui/components/_.web.ts
@@ -1,1 +1,0 @@
-export * from './web';
