+++ react/features/device-selection/actions.web.ts
