+++ react/features/participants-pane/components/native/ParticipantsPaneFooter.tsx
@@ -1,3 +1,4 @@
+/* eslint-disable lines-around-comment */
 import React, { useCallback } from 'react';
 import { View } from 'react-native';
 import { useDispatch, useSelector } from 'react-redux';
@@ -7,10 +8,13 @@ import { IconHorizontalPoints } from '../../../base/icons/svg';
 import Button from '../../../base/ui/components/native/Button';
 import IconButton from '../../../base/ui/components/native/IconButton';
 import { BUTTON_TYPES } from '../../../base/ui/constants';
+// @ts-ignore
 import MuteEveryoneDialog from '../../../video-menu/components/native/MuteEveryoneDialog';
 import { isMoreActionsVisible, isMuteAllVisible } from '../../functions';
 
+// @ts-ignore
 import { ContextMenuMore } from './ContextMenuMore';
+// @ts-ignore
 import styles from './styles';
