+++ react/features/settings/components/web/CalendarTab.js
@@ -5,7 +5,7 @@ import React, { Component } from 'react';
 
 import { translate } from '../../../base/i18n';
 import { connect } from '../../../base/redux';
-import { Button } from '../../../base/ui/components/web';
+import Button from '../../../base/ui/components/web/Button';
 import {
     CALENDAR_TYPE,
     MicrosoftSignInButton,
