+++ react/features/settings/functions.native.ts
@@ -1,0 +1,1 @@
+export * from './functions.any';
